misc/HtmlTestSuite/results/test-1096.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<link href="mailto:mstensho@chromium.org" rel="author" title="Morten Stenshorne"/>
<link href="https://html.spec.whatwg.org/multipage/rendering.html#frames-and-framesets" rel="help"/>
<link href="https://bugs.chromium.org/p/chromium/issues/detail?id=1116832" rel="help"/>
<link href="https://bugs.chromium.org/p/chromium/issues/detail?id=1353277" rel="help"/>
<script>
window.onload = () =&gt; {
  test(() =&gt; {
    const frameSet = document.querySelector('frameset');
    const frames = document.querySelectorAll('frame');
    assert_less_than(frames[0].offsetHeight, frameSet.offsetHeight);
    assert_greater_than(frames[0].offsetHeight, frames[1].offsetHeight);
    assert_greater_than_equal(frames[1].offsetHeight, 0);
  }, 'A large relative value should not produce weird sizes.');
};
</script>
</head><frameset frameborder="0" rows="4294967227*,*">
  <frame src="resources/green.html"/>
  <frame src="resources/red.html"/>
</frameset>
</html>
